+++ champss/pipeline_batch_db/sps_pipeline/workflow.py
@@ -32,6 +32,10 @@ docker_swarm_finished_states = [
 
 perpetual_processing_services = ["processing-manager", "processing-cleanup"]
 
+# Sometimes a Docker Swarm task gets stuck in pending/running state
+# indefinitely for unknown reasons...
+task_timeout_seconds = 60 * 40  # 40 minutes
+
 
 def message_slack(
     slack_message,
@@ -50,20 +54,49 @@ def message_slack(
         log.info(error)
 
 
-def wait_for_no_tasks_in_states(states_to_wait_for_none):
+def save_container_logs(service):
+    try:
+        log_text = ""
+        log_generator = service.logs(
+            details=True,
+            stdout=True,
+            stderr=True,
+            follow=False,
+        )
+        for log_chunk in log_generator:
+            log_text += log_chunk.decode("utf-8").strip() + "\n"
+
+        path = f"/data/chime/sps/logs/services/{service.name}.log"
+        with open(path, "w") as file:
+            file.write(log_text)
+    except Exception as error:
+        log.info(f"Error dumping logs at {path}: {error} (will skip gracefully).")
+
+
+def get_service_created_at_datetime(service):
+    try:
+        datetime = dt.datetime.strptime(
+            service.attrs["CreatedAt"].split(".")[0], "%Y-%m-%dT%H:%M:%S"
+        )
+        return datetime
+    except Exception as error:
+        log.info(
+            f"Error parsing CreatedAt for service {service}: {error} (will skip"
+            " gracefully)."
+        )
+        return None
+
+
+def wait_for_no_tasks_in_states(states_to_wait_for_none, docker_service_name_prefix=""):
     log.setLevel(logging.INFO)
 
     docker_client = docker.from_env()
-
-    # Sometimes a Docker Swarm task gets stuck in pending/running state
-    # indefinitely for unknown reasons...
-    task_timeout_seconds = 60 * 35  # 35 minutes from now
 
     is_task_in_state = True
 
     # For pending:
     # Docker Swarm can be freeze if too many tasks are in pending state concurrently.
-    # Additionally, Docker Swarm deqeues pending jobs in random order, so we need to
+    # Additionally, Docker Swarm dequeues pending jobs in random order, so we need to
     # only have one pending job at a time, to maintain ordering
     # For running:
     # It's helpful for Slack messages to wait for all running processes to finish
@@ -72,27 +105,33 @@ def wait_for_no_tasks_in_states(states_to_wait_for_none):
         is_task_in_state = False
 
         # Re-fetch Docker Swarm Service states
-        processing_services = sorted(
-            [
-                service
-                for service in docker_client.services.list()
-                if "processing" in service.name
-                and service.name not in perpetual_processing_services
-            ],
-            # Sort from oldest to newest to find finished services to remove
-            key=lambda service: dt.datetime.strptime(
-                # datetime does not support more than 6 decimals in seconds field
-                service.attrs["CreatedAt"][:-5],
-                "%Y-%m-%dT%H:%M:%S.%f",
-            ),
-        )
+        try:
+            processing_services = sorted(
+                [
+                    service
+                    for service in docker_client.services.list()
+                    if "processing" in service.name
+                    and docker_service_name_prefix in service.name
+                    and service.name not in perpetual_processing_services
+                ],
+                # Sort from oldest to newest to find finished services to remove
+                key=get_service_created_at_datetime,
+            )
+            processing_services = [
+                service for service in processing_services if service is not None
+            ]
+        except Exception as error:
+            log.info(
+                f"Error fetching Docker Swarm services: {error}."
+                f" Will stop waiting for tasks in states {states_to_wait_for_none}."
+            )
+            return
 
         for service in processing_services:
-            service_created_at = dt.datetime.strptime(
-                # datetime does not support more than 6 decimals in seconds field
-                service.attrs["CreatedAt"][:-5],
-                "%Y-%m-%dT%H:%M:%S.%f",
-            )
+            service_created_at = get_service_created_at_datetime(service)
+
+            if service_created_at is None:
+                continue
 
             try:
                 for task in service.tasks():
@@ -104,50 +143,66 @@ def wait_for_no_tasks_in_states(states_to_wait_for_none):
                             f"Removing finished service {service.name} in state"
                             f" {task_state}."
                         )
+
                         try:
+                            save_container_logs(service)
                             service.remove()
                         except Exception as error:
                             log.info(
                                 f"Error removing service {service.name}: {error} (will"
                                 " skip gracefully)."
                             )
-                    else:
-                        if task_state in docker_swarm_running_states:
-                            if (
-                                dt.datetime.now() - service_created_at
-                            ).total_seconds() > task_timeout_seconds:
+                    elif task_state in docker_swarm_running_states:
+                        if (dt.datetime.now() - service_created_at).total_seconds() > (
+                            task_timeout_seconds * 2
+                        ):
+                            log.info(
+                                f"Service {service.name} has been ruuning for more than"
+                                f" {(task_timeout_seconds  * 2) / 60} minutes in state"
+                                f" {task_state}, implying frozen task on 1st or 2nd"
+                                " final Workflow runner attempt. Will remove service."
+                            )
+
+                            try:
+                                save_container_logs(service)
+                                service.remove()
+                            except Exception as error:
                                 log.info(
-                                    "Approaching Workflow timeout for stuck service"
-                                    f" {service.name},"
-                                    f" {task_timeout_seconds / 60} minutes in state"
-                                    f" {task_state} (if reached, will try once more)."
+                                    f"Error removing service {service.name}:"
+                                    f" {error} (will skip gracefully)."
                                 )
-                            elif states_to_wait_for_none == docker_swarm_running_states:
-                                is_task_in_state = True
-                                break
-                        elif task_state in docker_swarm_pending_states:
-                            if (
-                                dt.datetime.now() - service_created_at
-                            ).total_seconds() > task_timeout_seconds:
-                                message_slack(
-                                    f"Removing stuck task {service.name}:{task_id},"
-                                    f" {task_timeout_seconds / 60} minutes in state"
-                                    f" {task_state}."
+                        # Task in state running, and we want to wait for no running states
+                        # Loop will continue
+                        elif states_to_wait_for_none == docker_swarm_running_states:
+                            is_task_in_state = True
+                            break
+                    elif task_state in docker_swarm_pending_states:
+                        if (dt.datetime.now() - service_created_at).total_seconds() > (
+                            task_timeout_seconds * 2
+                        ):
+                            log.info(
+                                f"Service {service.name} has been pending for more than"
+                                f" {(task_timeout_seconds  * 2) / 60} minutes in state"
+                                f" {task_state}, implying failed Docker task"
+                                " scheduling. Will remove service."
+                            )
+
+                            try:
+                                save_container_logs(service)
+                                service.remove()
+                            except Exception as error:
+                                log.info(
+                                    f"Error removing service {service.name}:"
+                                    f" {error} (will skip gracefully)."
                                 )
-                                try:
-                                    service.remove()
-                                except Exception as error:
-                                    message_slack(
-                                        f"Error removing service {service.name} with"
-                                        f" stuck task {task_id}: {error} (will skip"
-                                        " gracefully)."
-                                    )
-                            elif states_to_wait_for_none == docker_swarm_pending_states:
-                                is_task_in_state = True
-                                break
+                        # Task in state pending, and we want to wait for no pending states
+                        # Loop will continue
+                        elif states_to_wait_for_none == docker_swarm_pending_states:
+                            is_task_in_state = True
+                            break
             except Exception as error:
                 log.info(
-                    f"Error checking tasks for service {service.name}: {error} (will"
+                    f"Error checking tasks for service {service}: {error} (will"
                     " skip gracefully)."
                 )
 
@@ -161,7 +216,7 @@ def schedule_workflow_job(
     docker_name,
     docker_memory_reservation,
     docker_password,
-    workflow_name,
+    workflow_buckets_name,
     workflow_function,
     workflow_params,
     workflow_tags,
@@ -171,89 +226,137 @@ def schedule_workflow_job(
 
     docker_client = docker.from_env()
 
-    docker_client.login(username="chimefrb", password=docker_password)
+    # This is needed to spawn a Docker Sevice if the Docker Image is on
+    # a private registry (e.g. ou chimefrb DockerHub account)
+    try:
+        docker_password_secret_name = "DOCKER_PASSWORD"
+        docker_password_secret_id = docker_client.secrets.list(
+            filters={"name": docker_password_secret_name}
+        )[0].id
+        docker_client.login(username="chimefrb", password=docker_password)
+    except Exception as error:
+        log.info(
+            f"Failed to login to DockerHub: {error}. "
+            f"Will try to schedule this task anyway."
+        )
 
     workflow_site = "chime"
     workflow_user = "CHAMPSS"
 
-    work = Work(pipeline=workflow_name, site=workflow_site, user=workflow_user)
-    work.function = workflow_function
-    work.parameters = workflow_params
-    work.tags = workflow_tags
-    work.config.archive.results = True
-    work.config.archive.plots = "pass"
-    work.config.archive.products = "pass"
-    work.retries = 1
-    work.timeout = 60 * 40  # 40 minutes timeout
-
-    wait_for_no_tasks_in_states(docker_swarm_pending_states)
-
-    work_id = work.deposit(return_ids=True)
-
-    docker_volumes = [
-        docker.types.Mount(
-            # Only way I know of to add custom shared memory size allocations with Docker Swarm
-            target="/dev/shm",
-            source="",  # Source value must be empty for tmpfs mounts
-            type="tmpfs",
-            tmpfs_size=int(
-                100 * 1e9
-            ),  # Just give it 100GB of a shared memory upper-limit
+    try:
+        work = Work(
+            pipeline=workflow_buckets_name, site=workflow_site, user=workflow_user
         )
-    ]
+        work.function = workflow_function
+        work.parameters = workflow_params
+        work.tags = workflow_tags
+        work.config.archive.results = True
+        work.config.archive.plots = "bypass"
+        work.config.archive.products = "bypass"
+        work.retries = 1
+        work.timeout = task_timeout_seconds
 
-    for mount_path in docker_mounts:
-        mount_paths = mount_path.split(":")
-        mount_source = mount_paths[0]
-        mount_target = mount_paths[1]
-        docker_volumes.append(
-            docker.types.Mount(target=mount_target, source=mount_source, type="bind")
+        wait_for_no_tasks_in_states(docker_swarm_pending_states)
+
+        work_id = work.deposit(return_ids=True)
+
+        docker_volumes = [
+            docker.types.Mount(
+                # Bind mount the Docker socket to allow Docker-in-Docker (Workflow-in-Workflow) usage
+                target="/var/run/docker.sock",
+                source="/var/run/docker.sock",
+                type="bind",
+            ),
+            docker.types.Mount(
+                # Only way I know of to add custom shared memory size allocations with Docker Swarm
+                target="/dev/shm",
+                source="",  # Source value must be empty for tmpfs mounts
+                type="tmpfs",
+                tmpfs_size=int(
+                    100 * 1e9
+                ),  # Just give it 100GB of a shared memory upper-limit
+            ),
+        ]
+
+        for mount_path in docker_mounts:
+            mount_paths = mount_path.split(":")
+            mount_source = mount_paths[0]
+            mount_target = mount_paths[1]
+            docker_volumes.append(
+                docker.types.Mount(
+                    target=mount_target, source=mount_source, type="bind"
+                )
+            )
+
+        docker_service = {
+            "image": docker_image,
+            # Can't have dots or slashes in Docker Service names
+            # All Docker Services made with this function will be prefixed with "processing-"
+            "name": f"processing-{docker_name.replace('.', '_').replace('/', '')}",
+            # Use one-shot Workflow runners since we need a new container per process for unique memory reservations
+            # (we currently only use Workflow as a wrapper for its additional features, e.g. frontend)
+            "command": (
+                "workflow run"
+                f" {workflow_buckets_name} {' '.join([f'--tag {tag}' for tag in workflow_tags])} --site"
+                f" {workflow_site} --lifetime 1 --sleep-time 0"
+            ),
+            # Using template Docker variables as in-container environment variables
+            # that allow us this access out-of-container information
+            "env": ["CONTAINER_NAME={{.Task.Name}}", "NODE_NAME={{.Node.Hostname}}"],
+            # This is neccessary to allow Pyroscope (py-spy) to work in Docker
+            # 'cap_add': ['SYS_PTRACE'],
+            # Again, using one-shot Docker Service tasks too
+            "mode": docker.types.ServiceMode("replicated", replicas=1),
+            "restart_policy": docker.types.RestartPolicy(
+                condition="none", max_attempts=0
+            ),
+            # Labels allow for easy filtering with Docker CLI
+            "labels": {"type": "processing"},
+            # The labels on the Docker Nodes are pre-empetively set beforehand
+            "constraints": ["node.labels.compute == true"],
+            # Must be in bytes
+            "resources": docker.types.Resources(
+                mem_reservation=int(docker_memory_reservation * 1e9)
+            ),
+            # Will throw an error if you give two of the same bind mount paths
+            # e.g. avoid double-mounting basepath and stackpath when they are the same
+            "mounts": docker_volumes,
+            # An externally created Docker Network that allows these spawned containers
+            # to communicate with other containers (MongoDB, Prometheus, etc) that are
+            # also manually added to this network
+            "networks": ["pipeline-network"],
+            # Secrets are put into /run/secrets/<secret_name> inside the container
+            # (in the case that a Workflow runner will also spawn Workflow runners)
+            "secrets": [
+                docker.types.SecretReference(
+                    docker_password_secret_id, docker_password_secret_name
+                )
+            ],
+        }
+
+        log.info(f"Creating Docker Service: \n{docker_service}")
+
+        # Wait a few seconds because Work might still not have propogated to Buckets
+        # and Workflow runner can pickup nothing and quietly exit
+        time.sleep(2)
+
+        docker_client.services.create(**docker_service)
+
+        wait_for_no_tasks_in_states(docker_swarm_pending_states)
+
+        return work_id[0]
+    except Exception as error:
+        log.info(
+            f"Failed to deposit Work or create Docker Service: {error}. "
+            f"Will not schedule this task."
         )
 
-    docker_service = {
-        "image": docker_image,
-        # Can't have dots or slashes in Docker Service names
-        "name": f"processing-{docker_name.replace('.', '_').replace('/', '')}",
-        # Use one-shot Workflow runners since we need a new container per process for unique memory reservations
-        # (we currently only use Workflow as a wrapper for its additional features, e.g. frontend)
-        "command": (
-            "workflow run"
-            f" {workflow_name} {' '.join([f'--tag {tag}' for tag in workflow_tags])} --site"
-            f" {workflow_site} --lifetime 1 --sleep-time 0"
-        ),
-        # Using template Docker variables as in-container environment variables
-        # that allow us this access out-of-container information
-        "env": ["CONTAINER_NAME={{.Task.Name}}", "NODE_NAME={{.Node.Hostname}}"],
-        # This is neccessary to allow Pyroscope (py-spy) to work in Docker
-        # 'cap_add': ['SYS_PTRACE'],
-        # Again, using one-shot Docker Service tasks too
-        "mode": docker.types.ServiceMode("replicated", replicas=1),
-        "restart_policy": docker.types.RestartPolicy(condition="none", max_attempts=0),
-        # Labels allow for easy filtering with Docker CLI
-        "labels": {"type": "processing"},
-        # The labels on the Docker Nodes are pre-empetively set beforehand
-        "constraints": ["node.labels.compute == true"],
-        # Must be in bytes
-        "resources": docker.types.Resources(
-            mem_reservation=int(docker_memory_reservation * 1e9)
-        ),
-        # Will throw an error if you give two of the same bind mount paths
-        # e.g. avoid double-mounting basepath and stackpath when they are the same
-        "mounts": docker_volumes,
-        # An externally created Docker Network that allows these spawned containers
-        # to communicate with other containers (MongoDB, Prometheus, etc) that are
-        # also manually added to this network
-        "networks": ["pipeline-network"],
-    }
+        try:
+            work.delete()
+        except Exception as error:
+            log.info(f"Failed to delete dangling Work: {error}.")
 
-    log.info(f"Creating Docker Service: \n{docker_service}")
-
-    docker_client.services.create(**docker_service)
-
-    # Wait a second before querying Docker Swarm again
-    time.sleep(1)
-
-    return work_id[0]
+        return ""
 
 
 @click.command()
@@ -266,19 +369,23 @@ def schedule_workflow_job(
 def clear_workflow_buckets(workflow_buckets_name):
     """Function to empty given SPS Buckets collection on-site."""
     try:
+        print("test")
         buckets_api = Buckets()
+        print(buckets_api)
         # Bucket API only allows 100 deletes per request
         buckets_list = buckets_api.view(
             query={"pipeline": workflow_buckets_name},
             limit=100,
             projection={"id": True},
         )
+        print(buckets_list)
         while len(buckets_list) != 0:
             buckets_list = buckets_api.view(
                 query={"pipeline": workflow_buckets_name},
                 limit=100,
                 projection={"id": True},
             )
+            print(buckets_list)
             bucket_ids_to_delete = [bucket["id"] for bucket in buckets_list]
             log.info(f"Will delete buckets entries with ids: {bucket_ids_to_delete}")
             buckets_api.delete_ids(ids=bucket_ids_to_delete)
@@ -288,7 +395,7 @@ def clear_workflow_buckets(workflow_buckets_name):
                 projection={"id": True},
             )
     except Exception as error:
-        pass
+        print(error)
 
 
 @click.command()