+++ champss/sps-ops/sps_ops/__init__.py
@@ -1,1 +1,0 @@
-"""Module to store CHIME/SPS Operative code (Docker, Services, Ansible, etc.)"""